+++ src/machine/board_thingplus_rp2040.go
@@ -2,11 +2,6 @@
 
 package machine
 
-import (
-	"device/rp"
-	"runtime/interrupt"
-)
-
 // Onboard crystal oscillator frequency, in MHz.
 const xoscFreq = 12 // MHz
 
@@ -90,20 +85,7 @@ const (
 	UART_RX_PIN  = UART0_RX_PIN
 )
 
-// UART on the RP2040
-var (
-	UART0  = &_UART0
-	_UART0 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    rp.UART0,
-	}
-)
-
 var DefaultUART = UART0
-
-func init() {
-	UART0.Interrupt = interrupt.New(rp.IRQ_UART0_IRQ, _UART0.handleInterrupt)
-}
 
 // USB identifiers
 const (
