+++ src/machine/board_waveshare-rp2040-zero.go
@@ -7,11 +7,6 @@
 // - https://www.waveshare.com/wiki/RP2040-Zero
 package machine
 
-import (
-	"device/rp"
-	"runtime/interrupt"
-)
-
 // Digital Pins
 const (
 	D0  Pin = GPIO0
@@ -94,26 +89,7 @@ const (
 	UART1_RX_PIN = GPIO9
 )
 
-// UART on the RP2040
-var (
-	UART0  = &_UART0
-	_UART0 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    rp.UART0,
-	}
-	UART1  = &_UART1
-	_UART1 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    rp.UART1,
-	}
-)
-
 var DefaultUART = UART0
-
-func init() {
-	UART0.Interrupt = interrupt.New(rp.IRQ_UART0_IRQ, _UART0.handleInterrupt)
-	UART1.Interrupt = interrupt.New(rp.IRQ_UART1_IRQ, _UART1.handleInterrupt)
-}
 
 // USB CDC identifiers
 const (
