+++ src/machine/board_xiao-rp2040.go
@@ -7,11 +7,6 @@
 // - https://wiki.seeedstudio.com/XIAO-RP2040/
 package machine
 
-import (
-	"device/rp"
-	"runtime/interrupt"
-)
-
 // Digital Pins
 const (
 	D0  Pin = GPIO26
@@ -81,20 +76,7 @@ const (
 	UART_RX_PIN  = UART0_RX_PIN
 )
 
-// UART on the RP2040
-var (
-	UART0  = &_UART0
-	_UART0 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    rp.UART0,
-	}
-)
-
 var DefaultUART = UART0
-
-func init() {
-	UART0.Interrupt = interrupt.New(rp.IRQ_UART0_IRQ, _UART0.handleInterrupt)
-}
 
 // USB CDC identifiers
 const (
