+++ src/machine/machine_rp2040_gpio.go
@@ -331,3 +331,23 @@ func (p Pin) ioIntBit(change PinChange) uint32 {
 func getIntChange(p Pin, status uint32) PinChange {
 	return PinChange(status>>(4*(p%8))) & 0xf
 }
+
+// UART on the RP2040
+var (
+	UART0  = &_UART0
+	_UART0 = UART{
+		Buffer: NewRingBuffer(),
+		Bus:    rp.UART0,
+	}
+
+	UART1  = &_UART1
+	_UART1 = UART{
+		Buffer: NewRingBuffer(),
+		Bus:    rp.UART1,
+	}
+)
+
+func init() {
+	UART0.Interrupt = interrupt.New(rp.IRQ_UART0_IRQ, _UART0.handleInterrupt)
+	UART1.Interrupt = interrupt.New(rp.IRQ_UART1_IRQ, _UART1.handleInterrupt)
+}
